requirements.txt
# 基本
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0

# BASE自動出品
playwright>=1.40.0

# デザイン生成（Gemini）
google-genai>=1.0.0

# ブログ記事パイプライン（WordPress連携）
# 上記の requests で対応済み
